configurable hiding of notification icon, working delete of filtered calls from history

+++ DNDCallBlocker/src/hu/zoliweb/android/dndcallblocker/DNDCallBlockerIntentService.java
@@ -23,7 +23,6 @@
 package hu.zoliweb.android.dndcallblocker;
 
 import java.lang.reflect.Method;
-import java.util.ArrayList;
 
 import android.app.IntentService;
 import android.content.Context;
@@ -39,8 +38,9 @@ import com.android.internal.telephony.ITelephony;
 
 public class DNDCallBlockerIntentService extends IntentService {
 	private static final String DNDTAG = "DNDCallBlocker";
+	private SharedPreferences _prefs;
 	private String _phoneNr;
-	
+
 	DNDCallBlockerDBAdapter logDBAdapter;
 
 	public DNDCallBlockerIntentService() {
@@ -51,8 +51,10 @@ public class DNDCallBlockerIntentService extends IntentService {
 	protected void onHandleIntent(Intent intent) {
 		Context context = getBaseContext();
 
+		_prefs = PreferenceManager.getDefaultSharedPreferences(context);
+
 		Log.d(DNDTAG, "SRV: Got control.");
-		
+
 		_phoneNr = intent.getStringExtra("phone_nr");
 
 		// Make sure the phone is still ringing
@@ -70,25 +72,38 @@ public class DNDCallBlockerIntentService extends IntentService {
 			e.printStackTrace();
 			Log.d(DNDTAG, "Error trying to reject using telephony service.");
 		}
-		
+
 		// save event in call filter log
 		logDBAdapter = new DNDCallBlockerDBAdapter(context);
 		logDBAdapter.open();
-		if (_phoneNr != null)
-		{
+		if (_phoneNr != null) {
 			logDBAdapter.insertToLog(_phoneNr.trim());
 		} else {
 			logDBAdapter.insertToLog("");
 		}
-        logDBAdapter.close();
-		
-		// wait 2 sec, then delete last call from phone history
-		try {
-			Thread.sleep(2000);
-			deleteLastCall(context);
-		} catch (Exception e) {
-			e.printStackTrace();
-			Log.d(DNDTAG, "There was some problem with deleteLastCall procedure.");
+		logDBAdapter.close();
+
+		// delete last call from phone history
+		Boolean delete_history = _prefs.getBoolean("delete_history", true);
+		if (delete_history) {
+			Log.d(DNDTAG, "Trying to delete from history...");
+			try {
+				/*
+				int loopCnt = 0;
+				while (!deleteLastCall(context) && loopCnt < 16) {
+					loopCnt++;
+					Thread.sleep(125);
+				}
+				Log.d(DNDTAG, "History delete loop is over. cnt=" + loopCnt);
+				*/
+				Thread.sleep(2000);
+				deleteLastCall(context);
+				Log.d(DNDTAG, "History delete is over.");
+			} catch (Exception e) {
+				e.printStackTrace();
+				Log.d(DNDTAG,
+						"There was some problem with deleteLastCall procedure.");
+			}
 		}
 
 		return;
@@ -105,11 +120,8 @@ public class DNDCallBlockerIntentService extends IntentService {
 		ITelephony telephonyService;
 		telephonyService = (ITelephony) m.invoke(tm);
 
-		// Load preferences
-		SharedPreferences prefs = PreferenceManager
-				.getDefaultSharedPreferences(context);
 		// call handling preference
-		String handle_call = prefs.getString("handle_call", "silence");
+		String handle_call = _prefs.getString("handle_call", "silence");
 
 		Log.d(DNDTAG, "SRV: pref=".concat(handle_call));
 
@@ -118,7 +130,7 @@ public class DNDCallBlockerIntentService extends IntentService {
 				.getSystemService(Context.AUDIO_SERVICE);
 		// save original audio state
 		int old_mode = am.getRingerMode();
-		//set to silent
+		// set to silent
 		am.setRingerMode(AudioManager.RINGER_MODE_SILENT);
 
 		if (handle_call.equals("block")) {
@@ -155,25 +167,35 @@ public class DNDCallBlockerIntentService extends IntentService {
 	private boolean deleteLastCall(Context context) {
 		boolean isDeleted = false;
 		// Load the calls Log data via context calls Content resolver
-		//TODO: add selection to query
+		String extraQuerySelection = "";
+		// TODO: improve delete with date selection
+
+		if (_phoneNr != null) {
+			extraQuerySelection = " and " + android.provider.CallLog.Calls.NUMBER + " = '"
+					+ _phoneNr + "'";
+		} else {
+			Log.d(DNDTAG, "History delete... phone nr is null");
+			extraQuerySelection = " and " + android.provider.CallLog.Calls.NUMBER
+					+ " in ( '-1', '-2' )"; }
+		 
 		android.database.Cursor c = context.getContentResolver().query(
-				android.provider.CallLog.Calls.CONTENT_URI, null, null, null,
+				android.provider.CallLog.Calls.CONTENT_URI, null, 
+				android.provider.CallLog.Calls.TYPE + " in (" + android.provider.CallLog.Calls.INCOMING_TYPE + ", " + android.provider.CallLog.Calls.MISSED_TYPE + " )" + extraQuerySelection
+				, null,
 				android.provider.CallLog.Calls.DATE + " DESC");
-		ArrayList<CallRecord> CallRecordsList = new ArrayList<CallRecord>();
-		// Retrieve the column-indices of phoneNumber, date and calltype
+
+		String callRecordRowId = null;
+
 		Cursor cursor = c;
-		if (cursor.moveToFirst()) {
-			while (!cursor.isAfterLast()) {
-				// Load attributes of last call
-				CallRecord CallRecords = LoadCallRecord(cursor);
-				CallRecordsList.add(CallRecords);
-				cursor.moveToNext();
-			}
+		if (cursor.moveToFirst() && !cursor.isAfterLast()) {
+			// Load ID of last call
+			callRecordRowId = getLasCallRecordID(cursor);
 		}
-		if (CallRecordsList.size() > 0) {
+
+		if (callRecordRowId != null) {
 			int deletedRows = context.getContentResolver().delete(
 					android.provider.CallLog.Calls.CONTENT_URI,
-					"_ID=" + CallRecordsList.get(0).row_id, null);
+					"_ID=" + callRecordRowId, null);
 			if (deletedRows > 0)
 				isDeleted = true;
 			else
@@ -183,22 +205,8 @@ public class DNDCallBlockerIntentService extends IntentService {
 		return isDeleted;
 	}
 
-	private CallRecord LoadCallRecord(Cursor cursor) {
-		CallRecord contactData = new CallRecord();
-		//TODO: needs restructuring
-		String[] ColumnNames = cursor.getColumnNames();
-		for (int intLoop = 0; intLoop < ColumnNames.length; intLoop++) {
-			// Load id of Last call
-			if (android.provider.CallLog.Calls._ID
-					.compareTo(ColumnNames[intLoop]) == 0) {
-				contactData.row_id = cursor.getString(intLoop);
-			}
-		}
-
-		return contactData;
-	}
-
-	class CallRecord {
-		public String row_id;
+	private String getLasCallRecordID(Cursor cursor) {
+		int idIndex = cursor.getColumnIndex(android.provider.CallLog.Calls._ID);
+		return cursor.getString(idIndex);
 	}
 }
